address minor @eternoendless comments
